+++ jsontree/src/commonMain/kotlin/com/sebastianneubauer/jsontree/search/JsonTreeSearch.kt
@@ -1,5 +1,6 @@
-package com.sebastianneubauer.jsontree
+package com.sebastianneubauer.jsontree.search
 
+import com.sebastianneubauer.jsontree.JsonTreeElement
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
